Change height state type from Int to Double + add message for fetching door position

+++ lightwave-rooms/src/main/scala/de/lightwave/rooms/engine/mapping/RoomModelParser.scala
@@ -38,8 +38,8 @@ object RoomModelParser {
     map
   }
 
-  def toHeightMap(model: RoomModel): RoomMap[Int] = {
-    val map = toEmptyMap[Int](model.heightMap)
+  def toHeightMap(model: RoomModel): RoomMap[Double] = {
+    val map = toEmptyMap[Double](model.heightMap)
     val cols = model.heightMap.split("\\r?\\n")
 
     for {
